+++ zcash_client_backend/src/data_api/testing/pool.rs
@@ -998,7 +998,10 @@ pub fn spend_everything_multi_step_single_note_proposed_transfer<T: ShieldedPool
     Dsf: DataStoreFactory,
     <Dsf as DataStoreFactory>::AccountId: std::fmt::Debug,
 {
-    use crate::data_api::{MaxSpendMode, OutputOfSentTx, testing::transparent::GapLimits};
+    use crate::{
+        data_api::{MaxSpendMode, OutputOfSentTx},
+        wallet::transparent::GapLimits,
+    };
 
     let mut st = TestDsl::with_sapling_birthday_account(ds_factory, cache)
         .map(|builder| builder.with_gap_limits(GapLimits::new(10, 5, 3)))
@@ -1155,7 +1158,7 @@ pub fn spend_everything_multi_step_many_notes_proposed_transfer<T: ShieldedPoolT
     Dsf: DataStoreFactory,
     <Dsf as DataStoreFactory>::AccountId: std::fmt::Debug,
 {
-    use crate::data_api::{OutputOfSentTx, testing::transparent::GapLimits};
+    use crate::{data_api::OutputOfSentTx, wallet::transparent::GapLimits};
 
     let mut st = TestDsl::with_sapling_birthday_account(ds_factory, cache)
         .map(|builder| builder.with_gap_limits(GapLimits::new(10, 5, 3)))
@@ -1309,7 +1312,10 @@ pub fn spend_everything_multi_step_with_marginal_notes_proposed_transfer<
     Dsf: DataStoreFactory,
     <Dsf as DataStoreFactory>::AccountId: std::fmt::Debug,
 {
-    use crate::data_api::{MaxSpendMode, OutputOfSentTx, testing::transparent::GapLimits};
+    use crate::{
+        data_api::{MaxSpendMode, OutputOfSentTx},
+        wallet::transparent::GapLimits,
+    };
 
     let mut st = TestDsl::with_sapling_birthday_account(ds_factory, cache)
         .map(|builder| builder.with_gap_limits(GapLimits::new(10, 5, 3)))
@@ -1625,8 +1631,8 @@ pub fn send_multi_step_proposed_transfer<T: ShieldedPoolTester, Dsf>(
     Dsf: DataStoreFactory,
 {
     use crate::{
-        data_api::{OutputOfSentTx, TransactionStatus, testing::transparent::GapLimits},
-        wallet::{Exposure, TransparentAddressSource},
+        data_api::{OutputOfSentTx, TransactionStatus},
+        wallet::{Exposure, TransparentAddressSource, transparent::GapLimits},
     };
 
     let gap_limits = GapLimits::new(10, 5, 3);
@@ -2130,7 +2136,7 @@ pub fn spend_all_funds_multi_step_proposed_transfer<T: ShieldedPoolTester, Dsf>(
     Dsf: DataStoreFactory,
     <Dsf as DataStoreFactory>::AccountId: std::fmt::Debug,
 {
-    use crate::data_api::{OutputOfSentTx, testing::transparent::GapLimits};
+    use crate::{data_api::OutputOfSentTx, wallet::transparent::GapLimits};
 
     let mut st = TestDsl::with_sapling_birthday_account(ds_factory, cache)
         .map(|builder| builder.with_gap_limits(GapLimits::new(10, 5, 3)))