+++ zcash_client_memory/src/testing/mod.rs
@@ -24,8 +24,8 @@ use crate::{Account, AccountId, Error, MemBlockCache, MemoryWalletDb, SentNoteId
 
 #[cfg(feature = "transparent-inputs")]
 use zcash_client_backend::{
-    data_api::{InputSource, WalletRead, testing::transparent::GapLimits, wallet::TargetHeight},
-    wallet::WalletTransparentOutput,
+    data_api::{InputSource, WalletRead, wallet::TargetHeight},
+    wallet::{WalletTransparentOutput, transparent::GapLimits},
 };
 
 pub mod pool;