+++ zcash_client_backend/CHANGELOG.md
@@ -10,6 +10,9 @@ workspace.
 
 ## [Unreleased]
 
+### Added
+- `zcash_client_backend::wallet::transparent`
+
 ### Changed
 - `zcash_client_backend::wallet::OvkPolicy` has been substantially modified
   to reflect the view that a single outgoing viewing key should be uniformly
@@ -31,6 +34,10 @@ workspace.
 - `zcash_client_backend::wallet::GapMetadata` has been moved behind the
   `transparent-inputs` feature flag, as it is only useful in the context
   of the wallet receiving transparent funds.
+
+### Removed
+- `zcash_client_backend::data_api::testing::transparent::GapLimits` use
+  `zcash_client_backend::wallet::transparent::GapLimits` instead.
 
 ## [0.21.0] - 2025-11-05